Added shitty notes, tested torus rotation and rigid bodies
rotation and other depends on Origin point. I can make textures and collision meshes for rigid bodies, but we'll have to adjust them manually in godot which sucks but oh well.
HOWEVER, if I understand it correctly, if the name and order structure remains the same, then updating the mesh of an object WILL NOT remove/unlink our modifcations. (tested just now with the 3rd abominable torus, I moved and edited its shape, and the rotation script remained attached. This is promising news.)